pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-f5tts-api"
version = "1.0.0"
description = "Django Web API for F5-TTS: Text-to-Speech Service"
readme = "README.md"
license = {text = "MIT License"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Framework :: Django",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
requires-python = ">=3.10"
dependencies = [
    "django>=4.2.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.0.0",
    "f5-tts>=0.2.1",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "librosa",
    "soundfile",
    "numpy<=1.26.4",
    "pydub",
    "uvicorn>=0.22.0",
    "fastapi>=0.100.0",
    "python-multipart",
    "gradio>=3.45.2",
    "accelerate>=0.33.0",
    "transformers",
    "safetensors",
    "cached_path",
    "tqdm>=4.65.0",
    "matplotlib",
    "vocos",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "black",
    "ruff",
    "isort",
    "mypy",
    "django-debug-toolbar",
]
production = [
    "gunicorn",
    "psycopg2-binary",
    "redis",
    "celery",
]

[project.urls]
Homepage = "https://github.com/YourUsername/django-f5tts-api"
Repository = "https://github.com/YourUsername/django-f5tts-api.git"

[tool.setuptools.packages.find]
where = ["."]
include = ["django_f5tts*", "tts_api*"]

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
    migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
force_single_line = true
lines_after_imports = 2
skip = ["migrations"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
exclude = [
    "migrations/",
    "venv/",
    "env/",
]
